% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busTrace.R
\docType{data}
\name{busTrace}
\alias{busTrace}
\title{Bus velocity at given Time}
\format{A data frame with columns:
\describe{
 \item{Onibus}{Bus' series name}
 \item{Linha}{Bus' line name}
 \item{Velocidade}{Bus' velocity at given time}
 \item{HoraSegundos}{Time of record, in seconds}
}}
\source{
/href{https://crawdad.org/coppe-ufrj/RioBuses/20180319/}
}
\usage{
data(busTrace)
}
\description{
Bus velocity at given Time
}
\keyword{datasets}
