% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreesDBH.R
\docType{data}
\name{TreesDBH}
\alias{TreesDBH}
\title{Diameter at breast height of brazilian eucalyptus}
\format{A data frame with columns:
\describe{
 \item{parcela}{The Parcel's number (A primary group). Each parcel contais many trees at the same age}
 \item{idade}{Trees' age at given parcel, in months}
 \item{idadearred}{Trees' rounded age at given parcel, in months}
 \item{dap}{Diameter at breast height, or DBH. It is the standard for measuring trees. DBH refers to the tree diameter measured at 4.5 feet above the ground.}
}}
\source{
still unknown
}
\usage{
data(TreesDBH)
}
\description{
Diameter at breast height of brazilian eucalyptus
}
\keyword{datasets}
